Project_2/requirements.txt
langchain
langchain_community
langchain-google-genai
python-dotenv
streamlit
requests
langchain_experimental
sentence-transformers
langchain_chroma
langchainhub
pypdf
rapidocr-onnxruntime